refactor: migrate pages to dynamic language routing structure
- Migrate all pages from static structure to dynamic [lang] routing
- Maintain existing functionality while adding language parameter support
- Preserve all existing components and layouts
- Enable language-specific URL structure (/vi/, /en/)
