+++ src/app/(main)/[lang]/about/about-section.tsx
