+++ src/app/(main)/[lang]/about/company-history-timeline.tsx
