+++ src/app/(main)/[lang]/about/core-values-carousel.tsx
