+++ src/app/(main)/[lang]/about/core-values.data.ts
