+++ src/app/(main)/[lang]/about/page.tsx
@@ -9,3 +9,4 @@ export default function AboutPage() {
 		</>
 	);
 }
+
