+++ src/app/(main)/[lang]/contact/contact-form/index.ts
