+++ src/app/(main)/[lang]/contact/page.tsx
@@ -23,3 +23,4 @@ async function ContactPage() {
 		</>
 	);
 }
+
