+++ src/app/(main)/[lang]/hero-section.tsx
