+++ src/app/(main)/[lang]/layout/footer-loading.tsx
