+++ src/app/(main)/[lang]/layout/footer.tsx
