+++ src/app/(main)/[lang]/page.tsx
@@ -1,6 +1,7 @@
 import { Suspense } from "react";
-import HomepageLoading from "src/app/(main)/homepage-loading";
+import HomepageLoading from "src/app/(main)/[lang]/homepage-loading";
 import Prefetch from "src/components/behaviors/prefetch";
+import type { Locale } from "src/lib/i18n/config";
 import AboutSection from "./about/about-section";
 import ContactSectionDatabase from "./contact/contact-section-database";
 import HeroSection from "./hero-section";
@@ -13,25 +14,27 @@ import ServicesSectionDatabase from "./services/services-section-database";
  * Products, Services, Projects, and Contact use database data
  * About section remains static as it's not covered by admin database store
  */
-export default async function HomePage() {
+export default async function HomePage({ params }: { params: Promise<{ lang: Locale }> }) {
+	const { lang: locale } = await params;
+
 	return (
 		<Suspense fallback={<HomepageLoading />}>
-			<HomePageContent />
+			<HomePageContent locale={locale} />
 		</Suspense>
 	);
 }
 
-async function HomePageContent() {
+async function HomePageContent({ locale }: { locale: Locale }) {
 	return (
 		<>
 			<HeroSection />
 
 			<AboutSection />
 
-			<Prefetch url="/products" />
-			<Prefetch url="/services" />
-			<Prefetch url="/projects" />
-			<Prefetch url="/contact" />
+			<Prefetch url={`/${locale}/products`} />
+			<Prefetch url={`/${locale}/services`} />
+			<Prefetch url={`/${locale}/projects`} />
+			<Prefetch url={`/${locale}/contact`} />
 
 			<ProductsSectionDatabase />
 			<ServicesSectionDatabase />
