+++ src/app/(main)/[lang]/privacy-policy/page.tsx
