+++ src/app/(main)/[lang]/products/products-page-loading.tsx
