+++ src/app/(main)/[lang]/projects/projects-page-loading.tsx
